Move SetupAccountSite components into their own folder.
