+++ assets/js/modules/adsense/components/setup/v2/SetupAccountSite/Ready.js
@@ -26,13 +26,13 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import ViewContextContext from '../../../../../components/Root/ViewContextContext';
-import { trackEvent } from '../../../../../util';
-import { MODULES_ADSENSE } from '../../../datastore/constants';
-import SetupAccountSiteUI from './common/SetupAccountSiteUI';
+import ViewContextContext from '../../../../../../components/Root/ViewContextContext';
+import { trackEvent } from '../../../../../../util';
+import { MODULES_ADSENSE } from '../../../../datastore/constants';
+import SetupAccountSiteUI from '../common/SetupAccountSiteUI';
 const { useSelect, useDispatch } = Data;
 
-export default function SetupAccountSiteReady( { site, finishSetup } ) {
+export default function Ready( { site, finishSetup } ) {
 	const [
 		acknowledgedDisabledAutoAds,
 		setAcknowledgedDisabledAutoAds,
